requirements.txt
# Core Framework
streamlit>=1.28.0
python-dotenv>=1.0.0

# LangChain and LLM Integration
langchain>=0.1.14
langchain-openai>=0.1.7
langchain-community>=0.0.34

# PDF Processing
PyPDF2>=3.0.0

# Vector Database and Embeddings
faiss-cpu>=1.7.4
sentence-transformers>=2.2.0
huggingface_hub>=0.23.0

# Groq API (for LLM)
groq>=0.4.0

# Optional: Alternative LLM providers
# openai>=1.0.0  # Uncomment if using OpenAI
# google-generativeai>=0.3.0  # Uncomment if using Gemini

# Development and Utilities
pydantic>=2.0.0
typing-extensions>=4.0.0 